Fold the schedule form into the launch panel as a fourth mode
Scheduling was a separate panel that duplicated ~70% of the launch form, so
every field change had to be made twice. Promote it to a fourth mode in the
launch card alongside New/Resume/Attach and extract the shared agent, working
directory, title, model, effort, permission, and CLI-passthrough inputs into a
useLaunchForm hook plus a LaunchFormFields component that both modes render.
The schedule list is no longer part of the creation form; it moves to a
management-only ScheduledSessionsPanel that renders nothing until a schedule
exists.
